+++ CHANGELOG.md
@@ -61,6 +61,7 @@ Ref: https://keepachangelog.com/en/1.0.0/
   *  Attribute module param migration [#1927](https://github.com/provenance-io/provenance/pull/1927)
   *  Marker module param migration [#1934](https://github.com/provenance-io/provenance/pull/1934)
   *  Metadata module param migration [#1932](https://github.com/provenance-io/provenance/pull/1932)
+  *  Msgfees module param migration [#1936](https://github.com/provenance-io/provenance/pull/1936)
 * Restore the hold module [#1930](https://github.com/provenance-io/provenance/pull/1930).
 * Restore gov-prop cli commands and fix next key decoding [#1930](https://github.com/provenance-io/provenance/pull/1930).
 * Switch to InputOutputCoinsProv for exchange transfers [#1930](https://github.com/provenance-io/provenance/pull/1930).